Fix issue #2: Replace javah with the -h option in javac, in Makefile.
Usefully, the generic javac Make rule is still applicable, because
javac -h doesn't make a JNI header for classes without native methods;
only the JLcall class gets a header file, not the JLtest class.

I also enhanced the Makefile so that the testclass phony target has
the appclass target as a dependency.